Pawn 클래스 리팩토링 및 Pawn 기본 생성자 테스트 코드 추가
