--- src/main/java/softeer2nd/Pawn.java
@@ -1,14 +1,0 @@
-package softeer2nd;
-
-public class Pawn {
-
-    private String color;
-
-    public Pawn(String color) {
-        this.color = color;
-    }
-
-    public String getColor() {
-        return color;
-    }
-}
