+++ src/main/java/softeer2nd/chess/pieces/Pawn.java
@@ -1,0 +1,21 @@
+package softeer2nd.chess.pieces;
+
+public class Pawn {
+
+    public static final String WHITE_COLOR = "white";
+    public static final String BLACK_COLOR = "black";
+
+    private String color;
+
+    public Pawn() {
+        this.color = WHITE_COLOR;
+    }
+
+    public Pawn(String color) {
+        this.color = color;
+    }
+
+    public String getColor() {
+        return color;
+    }
+}
