+++ src/test/java/softeer2nd/chess/pieces/PawnTest.java
@@ -1,0 +1,29 @@
+package softeer2nd.chess.pieces;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class PawnTest {
+
+    @Test
+    @DisplayName("흰색과 검정색 폰이 생성되어야 한다")
+    void create() {
+        verifyPawn(Pawn.WHITE_COLOR);
+        verifyPawn(Pawn.BLACK_COLOR);
+    }
+
+    void verifyPawn(final String color) {
+        Pawn pawn = new Pawn(color);
+        assertThat(pawn.getColor()).isEqualTo(color);
+    }
+
+    @Test
+    @DisplayName("기본 생성자로 폰이 생성되어야 한다")
+    void create_기본생성자() {
+        Pawn pawn = new Pawn();
+        assertEquals(Pawn.WHITE_COLOR, pawn.getColor());
+    }
+}
